term_pretty: Fix peeking problem
The pretty printer peek was causing the print-out to be one
character behind the input. The peek was causing it to wait
for the next token after "." for example.

This was not noticeable when feeding it complete inputs but
it showed when running it interactively.

Only peek when printing a balanced open character. This will
block for more input but that is reasonable.